Mid refactoring commit, more moving of cache handling to come
